// discord/utils/types.ts
export interface DiscordSlashCommandOption {
  type: number; // 1=SUB_COMMAND, 2=SUB_COMMAND_GROUP, 3=STRING, 4=INTEGER, 5=BOOLEAN, 6=USER, 7=CHANNEL, 8=ROLE, 9=MENTIONABLE, 10=NUMBER, 11=ATTACHMENT
  name: string;
  description: string;
  required?: boolean;
  choices?: DiscordSlashCommandChoice[];
  options?: DiscordSlashCommandOption[];
  channel_types?: number[];
  min_value?: number;
  max_value?: number;
  min_length?: number;
  max_length?: number;
  autocomplete?: boolean;
}

export interface DiscordSlashCommandChoice {
  name: string;
  value: string | number;
}

export interface DiscordSlashCommand {
  id: string;
  type?: number;
  application_id: string;
  guild_id?: string;
  name: string;
  description: string;
  options?: DiscordSlashCommandOption[];
  default_member_permissions?: string | null;
  dm_permission?: boolean;
  default_permission?: boolean;
  nsfw?: boolean;
  version: string;
}